refactor: split dashboard.ts into domain-specific query modules
